pyproject.toml
[tool.poetry]
name = "morsum"
version = "0.1.0"
description = ""
authors = ["Javier Domingo Cansino <javierdo1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
Flask-SQLAlchemy = "^2.5.1"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.24.2"
mysqlclient = "^2.0.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
black = "^20.8b1"
isort = "^5.8.0"
flake8 = "^3.9.0"

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
